ultimos ajustes

+++ Gestion_SI/usuarios/funciones_usuarios.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
+#TODO: validar x tipo de dato, ex: prohibir que grupos y usuarios sean solo numeros.
+#TODO: modularizar aun mas, por ejemplo para usr_existe, para convertir  strings con coma a arrays
 
 ingresar_usuario(){ #Cumple la funcion de añadir nuevo usuario + agregarlo a grupos.(alta + modificacion) 
 	read -p "Ingrese nombre de usuario: " _user
-	usr_existe=$(cat /etc/passwd | grep $_user | wc -l)
-	if [ $usr_existe -eq 1 ]; then
+	usr_existe=$(cat /etc/passwd | grep -w $_user | wc -l) #-w para exact match
+	if [ $usr_existe -ne 0 ]; then
 	clear
 	echo "El usuario $_user ya existe."
 	return 1
 	fi
 
-	echo "Si desea agregar al usuario "$_user" a uno o mas grupos, ingrese los grupos separados por coma a continuacion, caso contrario, deje vacío: "
+	echo "Si desea agregar al usuario "$_user" a uno o mas grupos, ingrese los grupos separados por coma a continuacion, caso contrario, deje vacío."
 	echo ""
 	read -p "Grupos a agregar: " _grupos
 
@@ -23,14 +25,14 @@ ingresar_usuario(){ #Cumple la funcion de añadir nuevo usuario + agregarlo a gr
 		IFS="," #indicar delimitador
 		for grupo in $_grupos
 		do
-			grupo_existe=$(grep "$grupo" /etc/group | wc -l)
+			grupo_existe=$(grep -w "$grupo" /etc/group | wc -l)
 			if [ $grupo_existe -eq 0 ]; then
 				grupos_que_no_existen+=",$grupo"
 			else
 				grupos_con_coma+=",$grupo"
 			fi	
 		done
-
+#substrings para sacar las comas del principio del string y del final/ ,grupo,grupo,()=> grupo,grupo
 		grupos_con_coma=${grupos_con_coma#","}
 		grupos_con_coma=${grupos_con_coma%","}
 		grupos_que_no_existen=${grupos_que_no_existen#","}
@@ -42,9 +44,15 @@ ingresar_usuario(){ #Cumple la funcion de añadir nuevo usuario + agregarlo a gr
 		
 		case "$_crear" in
 			s)
-				grupos_que_no_existen="$(echo "$_grupos_que_no_existen" | tr -d ' ')" #remover espacios
-				. grupos/funciones_grupos.sh
-				ingresar_grupo "$grupos_que_no_existen"
+				echo "$_crear"
+				grupos_que_no_existen=$(echo "$grupos_que_no_existen" | tr -d ' ') #remover espacios
+				IFS=","
+				echo "$grupos_que_no_existen"
+				for grupo in $grupos_que_no_existen
+				do
+					echo $grupo	
+					groupadd "$grupo"
+				done
 				grupos_con_coma+=",$grupos_que_no_existen"
 				;;
 
@@ -66,12 +74,20 @@ ingresar_usuario(){ #Cumple la funcion de añadir nuevo usuario + agregarlo a gr
 
 eliminar_usuario(){
 	read -p "Ingrese el nombre del usuario que desea eliminar: " _user
+	
+	usr_existe=$(cat /etc/passwd | grep -w $_user | wc -l)
+	if [ $usr_existe -eq 0 ]; then
+		clear
+		echo "El usuario $_user no existe."
+		return 1
+	fi
+	
 	read -p "Seguro que desea eliminar al usuario $_user? (s/n):" opt
 
 	case "$opt" in
 		s)
 			userdel -r -f $_user
-			echo "El usuario $_user eliminado con éxito."
+			echo "El usuario $_user fue eliminado con éxito."
 			;;
 
 		n)
@@ -85,13 +101,28 @@ eliminar_usuario(){
 
 cambiar_uid(){
 	read -p "Ingrese el nombre del usuario del que desea cambiar su uid: " _user
+	
+
+	usr_existe=$(cat /etc/passwd | grep -w $_user | wc -l)
+	if [ $usr_existe -eq 0 ]; then
+		clear
+		echo "El usuario $_user no existe."
+		return 1
+	fi
+	
 	echo "UID actual: $(grep $_user /etc/passwd | cut -d: -f3)"
 	read -p "Ingrese la nueva UID para $_user: " _uid_nueva
 	usermod -u $_uid_nueva $_user
 }
 
 cambiar_nombre_usuario(){
-	read -p "Ingrese el nombre del usuario que desea cambiar: " _user
+	read -p "Ingrese el nombre del usuario que desea cambiar: " _user	
+	usr_existe=$(cat /etc/passwd | grep -w $_user | wc -l)
+	if [ $usr_existe -eq 0 ]; then
+		clear
+		echo "El usuario $_user no existe."
+		return 1
+	fi
 	read -p "Ingrese el nuevo nombre para el usuario $_user: " _userNuevo
 	mv /home/$_user /home/$_userNuevo 
 	usermod -l $_userNuevo $_user 
@@ -100,6 +131,15 @@ cambiar_nombre_usuario(){
 
 cambiar_contrasena_usuario(){
 	read -p "Ingrese el nombre del usuario del cual desea cambiar su contraseña: " _user
+	
+
+	usr_existe=$(cat /etc/passwd | grep -w $_user | wc -l)
+	if [ $usr_existe -eq 0 ]; then
+		clear
+		echo "El usuario $_user no existe."
+		return 1
+	fi
+	
 	passwd $_user
 }
 
@@ -114,7 +154,7 @@ eliminar_usuario_de_grupo(){
 	do
 		grupo_existe=$(grep "$grupo" /etc/group | wc -l)
 		if [ $grupo_existe -eq 1 ]; then
-			usuario_en_grupo=$(grep "$grupo" /etc/group | grep "$_user" | wc -l )
+			usuario_en_grupo=$(grep -w "$grupo" /etc/group | grep -w "$_user" | wc -l )
 			if [ $usuario_en_grupo -eq 0 ]; then
 				echo "El usuario $_user no pertence al grupo $grupo"
 			else
